ccdpack: Fix some compiler warnings in ccdpack C code
Mainly consting and removing unused variables. Fixed some problems
with variable clashes in STARCALL macro and size_t vs int
issues.

+++ applications/ccdpack/main/tclndf.h
@@ -90,7 +90,7 @@
       struct {
          char *data;               /* The text of the FITS headers */
          int loaded;               /* Has the FITS array been loaded? */
-         int ncard;                /* Number of FITS cards present */
+         size_t ncard;             /* Number of FITS cards present */
          HDSLoc *loc;              /* Locator to NDF .MORE.FITS extension */
       } fits;                   /* Contents of FITS extension. */
    } Ndf1;
